+++ packages/core/textarea/src/__tests__/unit/server-side.tsx
@@ -6,8 +6,12 @@ import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 import * as ReactDOMServer from 'react-dom/server';
 
 test('TextArea server side rendering', async () => {
-  (await getExamplesFor('textarea')).forEach(examples => {
-    const Example = require(examples.filePath).default;
-    expect(() => ReactDOMServer.renderToString(<Example />)).not.toThrowError();
-  });
+  (await getExamplesFor('textarea')).forEach(
+    (examples: { filePath: string }) => {
+      const Example = require(examples.filePath).default;
+      expect(() =>
+        ReactDOMServer.renderToString(<Example />),
+      ).not.toThrowError();
+    },
+  );
 });