+++ packages/elements/emoji/src/__tests__/unit/api/_resource-spec-util.ts
@@ -1,4 +1,6 @@
+import { ServiceConfig } from '@atlaskit/util-service-support';
 import { EmojiResource } from '../../../api/EmojiResource';
+import { EmojiResponse } from '../../../types';
 
 /**
  * #initSiteEmojiResource throws error
@@ -8,7 +10,10 @@ export class ErrorEmojiResource extends EmojiResource {
     return this.activeLoaders;
   }
 
-  initSiteEmojiResource(emojiResponse, provider): Promise<void> {
+  initSiteEmojiResource(
+    emojiResponse: EmojiResponse,
+    provider: ServiceConfig,
+  ): Promise<void> {
     throw new Error('Cannot initSiteEmojiResource');
   }
 }