+++ packages/elements/emoji/src/components/picker/CategoryTracker.ts
@@ -1,5 +1,4 @@
 import { List as VirtualList } from 'react-virtualized/dist/commonjs/List';
-
 import { CategoryId } from './categories';
 
 /**
@@ -34,7 +33,7 @@ export default class CategoryTracker {
 
   findNearestCategoryAbove(
     startIndex: number,
-    list: VirtualList,
+    list?: VirtualList,
   ): CategoryId | undefined {
     const rows = Array.from(this.rowToCategory.keys()).sort((a, b) => a - b);
     if (rows.length === 0) {