+++ packages/search/global-search/src/__tests__/unit/common/QuickSearchContainer.test.tsx
@@ -67,6 +67,7 @@ describe('QuickSearchContainer', () => {
       expect.any(Number),
       expect.any(String),
       defaultProps.createAnalyticsEvent,
+      expect.any(Number),
     ]);
   };