+++ packages/search/global-search/src/components/common/QuickSearchContainer.tsx
@@ -158,11 +158,28 @@ export class QuickSearchContainer extends React.Component<Props, State> {
     }
   };
 
-  fireExperimentExposureEvent = async (searchSessionId: string) => {
-    const { createAnalyticsEvent, getAbTestData, logger } = this.props;
+  fetchAbTestData = async (searchSessionId: string) => {
+    const { getAbTestData } = this.props;
+    const startTime = performanceNow();
+    const abTest = await getAbTestData(searchSessionId);
+    const elapsedMs = performanceNow() - startTime;
+
+    return {
+      elapsedMs,
+      abTest,
+    };
+  };
+
+  fireExperimentExposureEvent = async (
+    searchSessionId: string,
+    abTestPromise: Promise<ABTest | undefined>,
+  ) => {
+    const { createAnalyticsEvent, logger } = this.props;
+
+    const abTest = await abTestPromise;
+
     if (createAnalyticsEvent) {
       try {
-        const abTest = await getAbTestData(searchSessionId);
         if (abTest) {
           fireExperimentExposureEvent(
             abTest,
@@ -180,6 +197,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
     searchSessionId,
     recentItems,
     requestStartTime?: number,
+    experimentRequestDurationMs?: number,
   ) => {
     const { createAnalyticsEvent, getDisplayedResults } = this.props;
     if (createAnalyticsEvent && getDisplayedResults) {
@@ -199,6 +217,7 @@ export class QuickSearchContainer extends React.Component<Props, State> {
         elapsedMs,
         searchSessionId,
         createAnalyticsEvent,
+        experimentRequestDurationMs,
       );
     }
   };
@@ -276,7 +295,11 @@ export class QuickSearchContainer extends React.Component<Props, State> {
       });
     }
 
-    this.fireExperimentExposureEvent(this.state.searchSessionId);
+    const abTestPromise = this.fetchAbTestData(this.state.searchSessionId);
+    this.fireExperimentExposureEvent(
+      this.state.searchSessionId,
+      abTestPromise.then(({ abTest }) => abTest),
+    );
 
     try {
       const { results } = await this.props.getRecentItems(
@@ -287,12 +310,15 @@ export class QuickSearchContainer extends React.Component<Props, State> {
           recentItems: results,
           isLoading: false,
         },
-        () =>
+        async () => {
+          const experimentRequestDurationMs = (await abTestPromise).elapsedMs;
           this.fireShownPreQueryEvent(
             this.state.searchSessionId,
             this.state.recentItems || {},
             startTime,
-          ),
+            experimentRequestDurationMs,
+          );
+        },
       );
     } catch (e) {
       this.props.logger.safeError(